+++ modules/backend/assets/js/october-min.js
@@ -997,7 +997,7 @@ this.hide()
 var indicator=$('<div class="loading-indicator"></div>')
 indicator.append($('<div></div>').text(this.options.text))
 indicator.append($('<span></span>'))
-if(this.options.opaque!==undefined&&this.options.opaque){indicator.addClass('is-opaque')}
+if(this.options.opaque!==undefined){indicator.addClass('is-opaque')}
 this.$el.prepend(indicator)
 this.$el.addClass('in-progress')
 this.tally++}